+++ src/main/java/com/example/meosasshu/common/file/File.java
@@ -1,4 +1,4 @@
-package com.example.meosasshu.file;
+package com.example.meosasshu.common.file;
 
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
