+++ src/main/java/com/example/meosasshu/common/file/FileRepository.java
@@ -1,4 +1,4 @@
-package com.example.meosasshu.file;
+package com.example.meosasshu.common.file;
 
 import org.springframework.data.jpa.repository.JpaRepository;
