+++ src/main/java/com/example/meosasshu/common/redis/RedisProperties.java
@@ -1,4 +1,4 @@
-package com.example.meosasshu.redis;
+package com.example.meosasshu.common.redis;
 
 import lombok.Getter;
 import lombok.Setter;
