+++ src/main/java/com/example/meosasshu/controller/AccountController.java
@@ -1,13 +1,30 @@
 package com.example.meosasshu.controller;
 
 
+import com.example.meosasshu.common.security.user.CurrentUser;
+import com.example.meosasshu.common.security.user.UserDetailsImpl;
+import com.example.meosasshu.dto.response.AddressResDTO;
+import com.example.meosasshu.service.AccountService;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.annotation.Secured;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
-@RequiredArgsConstructor
 @RequestMapping("/api/v1/account")
+@RequiredArgsConstructor
 public class AccountController {
+
+    final private AccountService accountService;
+    @Secured("ROLE_USER")
+    @GetMapping("/address")
+    public ResponseEntity<AddressResDTO> getAddress(@AuthenticationPrincipal UserDetailsImpl userDetails){
+        AddressResDTO dto = accountService.getAddress(userDetails.getAccount());
+        return ResponseEntity.ok(dto);
+    }
+
 
 }
